+++ lib/action/core/contract_validator.rb
@@ -57,6 +57,8 @@ module Action
           elsif type == :uuid
             value.is_a?(String) && value.match?(/\A[0-9a-f]{8}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{12}\z/i)
           else
+            next true if Action.config.env.test? && value.class.name.start_with?("RSpec::Mocks::")
+
             value.is_a?(type)
           end
         end