+++ spec/action/enqueueable/via_sidekiq_spec.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require "sidekiq/testing"
-require "action/core/enqueueable"
+require "action/enqueueable"
 require "fixtures/enqueueable"
 
-RSpec.describe Action::Enqueueable, type: :worker do
+RSpec.describe Action::Enqueueable::ViaSidekiq, type: :worker do
   context "Interactor" do
     subject { TestEnqueueableInteractor.enqueue(this: "this", that: "that") }
 
@@ -77,7 +77,7 @@ RSpec.describe Action::Enqueueable, type: :worker do
         expect do
           subject
         end.to raise_error(ArgumentError,
-                           "Cannot pass non-JSON-serializable objects to Sidekiq. Make sure all objects in the context are serializable (or respond to to_global_id).") # rubocop:disable Layout/LineLength
+                           "Cannot pass non-JSON-serializable objects to Sidekiq. Make sure all expected arguments are serializable (or respond to to_global_id).") # rubocop:disable Layout/LineLength
       end
     end
   end