+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## UNRELEASED
-* N/A
+* Support blank exposures for `Action::Result.ok`
 
 ## 0.1.0-alpha.2.4.1
 * [FEAT] Adds full suite of per-Axn callbacks: `on_exception`, `on_failure`, `on_error`, `on_success`